feat(fonts): improve providers
