--- packages/astro/src/assets/fonts/helpers.ts
@@ -1,5 +1,0 @@
-import type { FontProvider } from './types.js';
-
-export function defineFontProvider<TName extends string>(provider: FontProvider<TName>) {
-	return provider;
-}
