--- packages/astro/src/assets/fonts/providers/bunny.ts
@@ -1,13 +1,0 @@
-import { defineFontProvider } from '../helpers.js';
-import { providers } from 'unifont';
-
-type Provider = typeof providers.bunny;
-
-export function bunny() {
-	return defineFontProvider({
-		name: 'bunny',
-		entrypoint: 'astro/assets/fonts/providers/bunny',
-	});
-}
-
-export const provider: Provider = providers.bunny;
