+++ packages/astro/src/assets/fonts/providers/entrypoints/adobe.ts
@@ -1,0 +1,4 @@
+import { providers } from 'unifont';
+
+// Required type annotation because its options type is not exported
+export const provider: typeof providers.adobe = providers.adobe;
