+++ packages/astro/src/assets/fonts/providers/entrypoints/fontshare.ts
@@ -1,0 +1,3 @@
+import { providers } from 'unifont';
+
+export const provider = providers.fontshare;
