+++ packages/astro/src/assets/fonts/providers/entrypoints/google.ts
@@ -1,0 +1,4 @@
+import { providers } from 'unifont';
+
+// Required type annotation because its options type is not exported
+export const provider: typeof providers.google = providers.google;
