--- packages/astro/src/assets/fonts/providers/fontshare.ts
@@ -1,13 +1,0 @@
-import { defineFontProvider } from '../helpers.js';
-import { providers } from 'unifont';
-
-type Provider = typeof providers.fontshare;
-
-export function fontshare() {
-	return defineFontProvider({
-		name: 'fontshare',
-		entrypoint: 'astro/assets/fonts/providers/fontshare',
-	});
-}
-
-export const provider: Provider = providers.fontshare;
