+++ packages/astro/src/assets/fonts/providers/index.ts
@@ -1,9 +1,34 @@
-import { adobe } from './adobe.js';
-import { bunny } from './bunny.js';
-import { fontshare } from './fontshare.js';
-import { fontsource } from './fontsource.js';
+import type { providers } from 'unifont';
+import type { FontProvider } from '../types.js';
 
-// TODO: move this file so it's not publicly exported
+function adobe(config: Parameters<typeof providers.adobe>[0]) {
+	return defineFontProvider({
+		name: 'adobe',
+		entrypoint: 'astro/assets/fonts/providers/adobe',
+		config,
+	});
+}
+
+function bunny() {
+	return defineFontProvider({
+		name: 'bunny',
+		entrypoint: 'astro/assets/fonts/providers/bunny',
+	});
+}
+
+function fontshare() {
+	return defineFontProvider({
+		name: 'fontshare',
+		entrypoint: 'astro/assets/fonts/providers/fontshare',
+	});
+}
+
+function fontsource() {
+	return defineFontProvider({
+		name: 'fontsource',
+		entrypoint: 'astro/assets/fonts/providers/fontsource',
+	});
+}
 
 /** TODO: jsdoc */
 export const fontProviders = {
@@ -12,3 +37,7 @@ export const fontProviders = {
 	fontshare,
 	fontsource,
 };
+
+export function defineFontProvider<TName extends string>(provider: FontProvider<TName>) {
+	return provider;
+}
